--Empty Message--

+++ blmarket/2013/facebookR2/A.cpp
@@ -30,7 +30,7 @@ int A[105];
 
 void process() 
 {
-    int ret = 1;
+    long long ret = 1;
     int prevline = 0;
     scanf("%d", &N);
     for(int i=0;i<N;i++) {
